requirements.txt
requests>=2.31.0
feedparser>=6.0.10
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
